refactor handle editor change to a separate file
